Fix POButtonStyle environments propagation
