+++ Sources/ProcessOutCoreUI/ProcessOutCoreUI.docc/ProcessOutCoreUI.md
@@ -14,6 +14,7 @@
 
 ### Button
 
+- ``POAnyButtonStyle``
 - ``POButtonStyle``
 - ``POButtonStateStyle``
 - ``SwiftUI/EnvironmentValues/isButtonLoading``
