+++ Sources/ProcessOutCoreUI/Sources/DesignSystem/Button/POAnyButtonStyle.swift
@@ -1,0 +1,25 @@
+//
+//  POAnyButtonStyle.swift
+//  ProcessOutCoreUI
+//
+//  Created by Andrii Vysotskyi on 20.10.2023.
+//
+
+import SwiftUI
+
+/// The `AnyHashable` type forwards body creation to an underlying style value,
+/// hiding the type of the wrapped value.
+public struct POAnyButtonStyle: ButtonStyle {
+
+    public init<Style: ButtonStyle>(erasing style: Style) {
+        self.style = style
+    }
+
+    public func makeBody(configuration: Configuration) -> AnyView {
+        AnyView(style.makeBody(configuration: configuration))
+    }
+
+    // MARK: - Private Properties
+
+    private let style: any ButtonStyle
+}
